backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
base58==2.1.1
certifi==2025.11.12
deprecated==1.3.1
dotenv==0.9.9
ecpy==1.2.5
fastapi==0.122.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-core==2.41.5
python-dotenv==1.2.1
qrcode==8.2
starlette==0.50.0
types-deprecated==1.3.1.20251101
typing-extensions==4.15.0
typing-inspection==0.4.2
websockets==15.0.1
wrapt==2.0.1
xrpl-py==4.3.1
